+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+[v0.2.0] 2024-10-19
+-------------------
+
+-   Remove makefile-lossless library
+-   PEG parser used for Makefile parsing
+-   Support multi target rules
+
 [v0.1.0] 2024-10-19
 -------------------